// mode-c/src/app/private/model/company.ts
export class Company{
    companyAddress:string | undefined;
    companyEmailAddress: string | undefined;
    companyId: string | undefined;
    companyName: string | undefined;
   
    emailAddress: string | undefined
    firstname: string | undefined;
    id: string | undefined;
    lastname: string | undefined;
    password: string | undefined;
    phoneNo: string | undefined;
    profilePicture: string | undefined;
    userName: string | undefined
  firstName: string | undefined;
  lastName: string | undefined;
  role: string | undefined;
    
}